Input etc: Fix small type issues on TextInput refs.
Resolving these `any`s into the proper type shows a couple of spots
where we should be checking for null.  Not sure if these cases can
actually be exercised to cause a live bug -- but this is the easiest
way to be sure.

Also remove the `MultilineInput#textInput` property, which doesn't
seem to be ever set; presumably left behind by a past refactor.