android notif: Handle `remove` events, just updating data.
These GCM messages will be sent by the Zulip server (planned for
1.9.1) when a Zulip message that caused a notification becomes read.

For a start, handle them just as far as removing this Zulip message
from the `conversations` data structure, which we use to generate the
full notification text when future messages come in.

Written in part pairing with @borisyankov.