webview: Set message node's `data-read` when scrolling
When scrolling the WebView, in addition to sending scroll events
to the outside world, iterate through scrolled message nodes and
mark them as read in the DOM tree.

Any message node already contains a `data-read`, `data-starred` etc.
based on the flags a given message has. Unread messages will be missing
the `data-read` attribute, so we set it to `true` if we scrolled pass them.