Rename WebView 'messages' entity to 'UpdateEvent'
`Message` is a very loaded term and we use it extensively for
Zulip's messages in a narrow. Anything different that is also
called a 'message' should probably be called something else.

Data coming out of the WebView are called 'MessageListEvent....'.
It only makes sense that the data coming into is also an event.

So, from now on, data coming to the WebView is `WebViewUpdateEvent`.

Rename all data types and data handlers to use that naming.