screen: Fix scroll in when `scrollEnabled={false}`
A change in how `Screen` component works when not scrollable
(026814edd), and a fix to not render the whole `FlatList`s and
`ScrollList`s (b0730a063), have revealed a bug -- the `ScrollView`
used inside the `Screen` is not allowing its content to be scrolled.

To fix that, adding `flex: 1` is enough.

Tested all screens with `ScrollEnabled={false}` on both iOS and Android.