message list: Reframe "render context" as "background data".
This gives it a somewhat clearer conceptual meaning, which will help
us reuse it in other parts of the message-list code outside of
rendering HTML, and eliminate the duplication with props.

Reflecting the new, clarified semantics, take `narrow` out of it; pass
that as a separate parameter in the handful of places it was needed.