title: Give TitlePrivate the specific user email, rather than `narrow`.
Writing `props.narrow[0].operand` was pretty grody; and
`getUserInPmNarrow(props.narrow)` was better because more explicit,
but both relied on an assumption that the narrow was necessarily a PM
narrow.  We do happen to know that from the parent component, but
there wasn't a way to see that locally.  Now there is.

And now `getUserInPmNarrow` is unused!  Delete it.