fetchActions: Delete bogus type FetchMessagesAction.
There's no actual action here that this type might refer to; the
actual type of the values we've been labelling with this type is a
thunk action, a function taking (Dispatch, GetState) and returning
sometimes void, sometimes Promise<void>.

Flow can infer that just fine if we leave it off, and because we
already have nice annotations on the parameters of the function,
there's nothing to be gained in clarity for a human reader either.
So, just cut it out.