emoji: Fix parsing of "keypad" and other multi-codepoint Unicode emoji.
Fixes #3129.

There are actually just 12 emoji like this so far: `:zero:` through
`:nine:`, and `:hash:` and `:asterisk:`.

There's still a bug that remains, which is that we think `:0:`
etc. are names for the 10 single-digit emoji (in addition to the
spelled-out `:zero:`), but the server seems to disagree; as a result
autocomplete and adding a reaction don't work.  But once there,
everything displays correctly.