flow: Re-enable warnings for unused fixmes.
I don't totally know why these fixmes are no longer needed; shrug.
They aren't the transitional ones that motivated us to disable
this temporarily, so apparently we successfully removed all of
those a while ago.

One of them is an issue that was *introduced* in the upgrade; note
there are a couple of others just like it that are still around.