tests: Fix check for running in Jest.
This check wasn't working reliably when running a subset of tests,
which is much more common now that we run it by default only on
changed files.

Instead, solve the problem directly, in the `.jest.globals` config.
This requires also a line each to give ESLint and Flow a heads-up that
this variable may exist and it's OK; and adjusting the update-libdefs
alias so it doesn't blow the Flow hint away.