i18n/messages: Re-enable lint, fixing extensions in imports.
ESLint was showing a pair of errors on each import, like these
(edited slightly for readability):

  error  Unable to resolve path to module './translations/messages_ar'
    import/no-unresolved
  error  Missing file extension for "./translations/messages_ar"
    import/extensions

Tell the `import` plugin that `.json` files are importable,
eliminating the `import/no-unresolved` errors; and take the
airbnb style's advice on the other errors by writing the
extensions explicitly.  I remember this pattern being pretty
confusing the first couple of times I saw it, so explicitly
saying it's a JSON file may help.

Leave the spellchecker disabled so we don't have to add a bunch of
"words" for these language codes.