tools/fmt: Move `yarn prettier` logic out of package-scripts, too.
This is a pure refactor.  There is one change where that isn't
evident: in the --all case, we pass `eslint` an argument of
`src/**/*.js` (the literal string, not the list of filenames it
might expand to), rather than `src/`.  The latter seems like a
much cleaner API, but `prettier` only takes the other one, and
it's clearer to match what we pass to the two of them.