tests: Downgrade spellcheck to a genuine warning level.
Now spellcheck warnings:
 (a) will not cause the test run to fail;
 (b) will not be printed at all in CI, or a default run
     of `tools/test --full`.

So they're information for you as you prepare a branch, but if you
decide a spellcheck warning is wrong, just ignore it unless it seems
likely to be common.

Hopefully this can let us stop maintaining so much of a "wordlist" of
one-off names, etc., that spellcheck complained about once.