login: Make `PasswordInput` styling more robust.
Fixes #3182.

This fixes an annoying issue that makes the size of the input
about half what it should be (on inspection 16px vs 33px for the
default one).

Digging through past commits I did not find the exact change that
caused this to break. The main cause though is that the styling was
a bit 'fragile'.

Fixing this turned out to be tricky. We need the height of the
wrapper `View` to be exactly the same as the `Input` but we can't
set it manually as it might vary - 33px on iOS and similar on Android
but also can change on zoom etc.

The fix:
 * does not try to change the Input sizing in any way
 * fixes the positioning of the 'Show' button

Tested on iOS and Android.