input: Handle full text/values localization after all.
In 6a10912c6 a couple of hours ago, I determined that all the call
sites of this component pass `string` for the `placeholder` prop, and
so reduced its type in order to better facilitate refactoring and
fixing of other types.

Turns out one call site was passing a value of type `LocalizableText`,
aka `any` :'-(, and so the type-checker didn't catch that I'd missed
that one.  I did grep through call sites... but that one was indirect,
and though I thought I'd hunted down the indirect call sites too,
evidently my grep wasn't thorough enough.  The result is that the app
crashes when you narrow to any conversation.  (Fortunately that's at
least a tough regression not to notice!)

Fix the regression by writing down a type that covers what ComposeBox
passes here via MultilineInput, and writing code to match.  It's
similar in its effect to what preceded 6a10912c6, but more explicit
because the latter is (quite reasonably) rejected by Flow once the
actual type involved is written down.

The root problem here is of course that nobody should have to grep
through caller's callers to work out what variety of values might show
up on a given line of code.  That is why we have a type-checker.  In
the next commit, I'll fix LocalizableText so that that type, at least,
never causes trouble again.