user-picker: Fix a completely bogus type, and its `any` enablers.
The `listRef` property on a `UserPickerCard` component is definitely
not a function!  It is (if not null or undefined) a component, and
indeed a particular type of component -- we rely on that in order to
call `scrollToEnd` on it.

Fix the types to reflect that; and now we get to mark these two files
strict-local.  Also make the props and state types exact, while here.