outbox: Drop sent messages when event queue re-established.
Fixes #3203.

We do not delete sent messages from outbox immediately on a
successful sending, but keep them in the outbox until we receive
the message from the server via our event queue.

This works great unless we never see the message as an event, either
because we didn't have a live event queue at the time we sent it or
because we lose the queue thereafter before we can drain it of events.

Both cases are fundamentally unavoidable.  The first case should be
rare but currently is more frequent than optimal.  The second case can
happen if the user quits the app, or goes offline, shortly after
sending, and doesn't come back within the few minutes that we
currently set the queue to last.

In any case, if we never get this event that happened, then one thing
that's sure to happen eventually (unless the user just stops using the
app) is that we establish a new queue.  Conversely, that means we've
given up on the old queue and are definitely never going to get that
old event from it.  Drop any sent messages when we see this.

We might decide to do this more aggressively in future but that might
also lead to behavior confusing to the user like "Where did my message
disappear?"

[greg: revised explanation in part]