types: Make toplevel-declared types exact, where no errors.
The situation isn't as uniform for all kinds of types in our code as
it is for React components' Props and State types (discussed in a
previous commit); but even in that broader context it's still true
that most object types should be exact, and when making them so
produces an error, that usually still seems to point to something
wrong in our code.  So, let's go switching object types to exact
where appropriate.

I edited all these and others by running this command (very similar
to the one in a preceding commit for Props and State types):

  perl -i -0pe \
    's/^((?:export )?type \S+ = \{)(\n.*?\n)?(\};)/$1|$2|$3/msg' \
    src/**/*.js

Then `yarn flow` found a ton of errors, as expected.

To make a start, I just used `git checkout @ ...` to discard all the
changes that caused errors; these are the ones that didn't.