android notif: Receive GCM messages with our own listener service.
This lets us cut out the wix code for another pathway.

The new code and manifest blobs are based on upstream GCM docs:
  https://developers.google.com/cloud-messaging/downstream
  https://developers.google.com/cloud-messaging/android/client

In the manifest we actually already had a partial blob of our own,
though it was erroneously still hardcoding `com.zulipmobile` rather
than varying appropriately for debug builds; delete that.  The wix
manifest differs by having an extra action in the GcmReceiver
intent-filter, for s/RECEIVE/REGISTRATION/; upstream docs mention that
in a box about if you want to support pre-KitKat versions of Android,
which we no longer do.

Like in the intent service in the parent commit, for now in the
service code we basically just inline what the wix code would do, to
keep changes small at this step.  Here that means we inline
GcmMessageHandlerService, then PushNotification.get.

We remove the `throws` because this doesn't actually throw, and
there's no point in littering the caller's code with a `try`/`catch`
that won't do anything.