android notif: Get those "talk to React" facades/helpers locally.
The two "helper" classes have no non-static data members at all -- the
actual objects are pointless, and the methods should just be static.

Moreover, we're going to use each of them at most once for a given
instance of GCMPushNotifications... so there's absolutely no reason
to allocate them up front and carry them around the whole class as
members.  Just allocate at the point where we need them, greatly
reducing the scope of code where they're visible.

The facade is a singleton (in its "holder"), and so there's no
reason we can't just `get` it at every point where we need it.
In notifyReact where we keep referring to it, it's convenient to
stick it in a local, though.

A lot of the methods on these are in turn trivial implementations
with complicated interfaces, like we've been seeing; dealing with
those will come later.